examples/requirements.txt
cryptography
paho-mqtt